ASL_AI/requirements.txt
# Core ML
tensorflow>=2.13.0
mediapipe>=0.10
scikit-learn>=1.3
opencv-python>=4.8
numpy>=1.24

# API
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
eventlet>=0.33.0

# Translation
deep-translator

# Unicode text rendering
Pillow>=10.0
